Ok enfin c'est bon, double précision (mais ca regle pas les problèmes)

+++ simulation/calc.h
@@ -3,7 +3,7 @@
 
 #include "utils.h"
 
-float floor_temp_calc(int i, int j, float lambda, float mu, float tau, surface_temp s_t, f_matrix* air_temp, float c_p, float m_i, float fluid_speed) {
+double floor_temp_calc(int i, int j, double lambda, double mu, double tau, surface_temp s_t, f_matrix* air_temp, double c_p, double m_i, double fluid_speed) {
     if (j > 0 && (j % air_temp->cols != 0)) {
         return (fluid_speed / 3.6 * air_temp[i].data[idx(air_temp[i].rows - 1, j - 1, air_temp[i].cols)] / lambda + air_temp[i].data[idx(air_temp[i].rows - 1, j, air_temp[i].cols)] / tau - (s_t.h / c_p) / m_i * (s_t.surf.length * s_t.surf.width) * s_t.temp) / (fluid_speed / 3.6 / lambda + 1 / tau - (s_t.h / c_p) * (s_t.surf.length * s_t.surf.width) / m_i);
     } else {
@@ -11,7 +11,7 @@ float floor_temp_calc(int i, int j, float lambda, float mu, float tau, surface_t
     }
 }
 
-float wall_temp_calc(int i, int j, int wall_idx, float lambda, float mu, float tau, surface_temp s_t, f_matrix* air_temp, float c_p, float m_i, float fluid_speed) {
+double wall_temp_calc(int i, int j, int wall_idx, double lambda, double mu, double tau, surface_temp s_t, f_matrix* air_temp, double c_p, double m_i, double fluid_speed) {
     if (j > 0 && (j % air_temp->cols != 0)) {
         return (fluid_speed / 3.6 * air_temp[wall_idx].data[idx(i, j - 1, air_temp[wall_idx].cols)] / lambda + air_temp[wall_idx].data[idx(i, j, air_temp[wall_idx].cols)] / tau - (s_t.h / c_p) / m_i * (s_t.surf.length * s_t.surf.width) * s_t.temp) / (fluid_speed / 3.6 / lambda + 1 / tau - (s_t.h / c_p) * (s_t.surf.length * s_t.surf.width) / m_i);
     } else {
@@ -19,9 +19,9 @@ float wall_temp_calc(int i, int j, int wall_idx, float lambda, float mu, float t
     }
 }
 
-float air_temp_calc(int x, int y, int z, float lambda, float mu, float h_n, int nb_subd, float tau, f_matrix* prev_temp, float D) {
+double air_temp_calc(int x, int y, int z, double lambda, double mu, double h_n, int nb_subd, double tau, f_matrix* prev_temp, double D) {
 
-    float temp_x_plus_1, temp_x_moins_1, temp_y_plus_1, temp_y_moins_1, temp_z_plus_1, temp_z_moins_1; // Représentent à l'itération précédente les températures décalées de + ou - 1 selon x, y ou z
+    double temp_x_plus_1, temp_x_moins_1, temp_y_plus_1, temp_y_moins_1, temp_z_plus_1, temp_z_moins_1; // Représentent à l'itération précédente les températures décalées de + ou - 1 selon x, y ou z
 
     assert( x > 0 && x < nb_subd - 1 && y > 0 && y < prev_temp[x].cols && z > 0 && z < prev_temp[x].rows);
     //if (x > 0)                 { temp_x_moins_1 = prev_temp[x-1].data[idx( z , y , prev_temp[x].cols)]; } else { temp_x_moins_1 = prev_temp[x].data[idx(z, y, prev_temp[x].cols)]; }
